test refactor accordingly to the convention
